+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - Added 'DEVELOPMENT' environment variable flag to disable shiritori rules for testing
+- Added lottery system
 
 ## Fixed